NILE-292:Moved cluster and dashboard services
