+++ src/app/pages/cluster/cluster-table-prime/cluster-table-prime.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
 import { ClusterEntity } from '../../../api/model/clusterEntity';
-import { ClusterTableService } from '../cluster-table/cluster-table.service';
+import { ClusterTableService } from '../../../services/cluster/cluster-table.service';
 import { ClusterWizardComponent } from '../cluster-wizard/cluster-wizard.component';
