+++ src/app/pages/cluster/cluster-table/cluster-table.component.ts
@@ -1,5 +1,5 @@
 import {Component, OnInit, ViewChild} from '@angular/core';
-import { ClusterTableService } from './cluster-table.service';
+import { ClusterTableService } from '../../../services/cluster/cluster-table.service';
 import { LocalDataSource } from 'ng2-smart-table';
 import { ClusterEntity } from '../../../api/model/clusterEntity';
 import {ClusterWizardComponent} from '../cluster-wizard/cluster-wizard.component';
