+++ src/app/pages/dashboard/dashboard.component.ts
@@ -1,5 +1,5 @@
 import { AfterViewInit, Component, OnInit } from '@angular/core';
-import { DashboardService } from './dashboard.service';
+import { DashboardService } from '../../services/dashboard/dashboard.service';
 
 @Component({
   selector: 'nile-dashboard',
