+++ src/app/pages/dashboard/dashboard.module.ts
@@ -4,9 +4,8 @@ import { CommonModule } from '@angular/common';
 import { NgaModule } from '../../theme/nga.module';
 
 import { DashboardComponent } from './dashboard.component';
-import { DashboardService } from './dashboard.service';
-import { ClusterSummaryChartComponent } from '../../components/cluster-summary-chart';
-import { ResourceSummaryChartComponent } from '../../components/resource-summary-chart';
+import { DashboardService } from '../../services/dashboard/dashboard.service';
+
 import { ComponentsModule } from '../../components/components.module';
 import { ProjectService } from '../../services/project';
