+++ src/app/services/cluster/index.ts
@@ -1,0 +1,1 @@
+export * from './cluster-table.service';
