+++ src/app/services/dashboard/index.ts
@@ -1,0 +1,1 @@
+export * from './dashboard.service';
