+++ src/app/services/index.ts
@@ -1,1 +1,3 @@
 export * from './project';
+export * from './dashboard';
+export * from './cluster';
